Expose WireBase. Split shader translator compilation units
