+++ src/extra/CMakeLists.txt
@@ -9,6 +9,9 @@ set(extra_SOURCES
   gfx/shader/translator.cpp
   gfx/shader/wgsl.cpp
   gfx/shader/composition.cpp
+  gfx/shader/linalg_shards.cpp
+  gfx/shader/math_shards.cpp
+  gfx/shader/core_shards.cpp
   gfx/gltf.cpp
   gfx/camera.cpp
   gfx/texture.cpp
