+++ src/extra/gfx/shader/translator.hpp
@@ -217,6 +217,8 @@ public:
   ITranslationHandler *resolve(Shard *shard);
 };
 
+TranslationRegistry &getTranslationRegistry();
+
 void applyShaderEntryPoint(SHContext *context, shader::EntryPoint &entryPoint, const SHVar &input);
 
 } // namespace shader
